revert to match pxt-arcade
