+++ libs/arcadeshield/_locales/arcadeshield-strings.json
@@ -55,7 +55,6 @@
   "{id:category}Math": "Math",
   "{id:category}Power": "Power",
   "{id:category}Scene": "Scene",
-  "{id:category}Screen": "Screen",
   "{id:category}ScreenBitmap": "ScreenBitmap",
   "{id:category}Texteffects": "Texteffects",
   "{id:category}_helpers_workaround": "_helpers_workaround",
