+++ libs/arcadeshield/context/console.ts
@@ -5,8 +5,8 @@ namespace scene.consoleOverlay {
     const marginx = 4;
     const marginy = 2;
     const consoleFont = bitmap.font5;
-    const consoleLines = Math.floor(screen.height / (consoleFont.charHeight + marginy)) - 1;
-    const consoleColumns = Math.floor((screen.width - 2 * marginx) / consoleFont.charWidth);
+    const consoleLines = Math.floor(theScreen.height / (consoleFont.charHeight + marginy)) - 1;
+    const consoleColumns = Math.floor((theScreen.width - 2 * marginx) / consoleFont.charWidth);
     console.addListener(listener);
 
     export function isVisible() {
