--- libs/arcadeshield/screen.ts
@@ -1,30 +1,0 @@
-namespace screen {
-    //% shim=pxt::updateScreen
-    export function updateScreen(img: Bitmap): void {
-        return
-    }
-    //% shim=pxt::updateStats
-    export function updateStats(s: string): void {
-        return
-    }
-        //% shim=pxt::displayHeight
-    export function displayHeight(): number {
-        return 0
-    }
-    //% shim=pxt::displayWidth
-    export function displayWidth(): number {
-        return 0
-    }
-    //% shim=pxt::displayPresent
-    export function displayPresent(): boolean {
-        return false
-    }
-    //% shim=pxt::setPalette
-    export function setPalette(buf: Buffer): void {
-        return
-    }
-    //% shim=pxt::setScreenBrightness
-    export function setScreenBrightness(b: number): void {
-        return
-    }
-}
